+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/response/DisplayResponseViewState.kt
@@ -19,4 +19,5 @@ data class DisplayResponseViewState(
     val canCopy: Boolean,
     val canSave: Boolean,
     val isSaving: Boolean = false,
+    val showExternalUrlWarning: Boolean,
 )