+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - When using multiple variables in a shortcut, their values are now resolved in deterministic order, according to the order in which the variables appear on the Variables screen
 - Built-in icons are now treated as adaptive icons, allowing them to look nicer on the home screen of devices that support this
 
+### Miscellaneous
+- When an HTML response tries to open a URL, either because of a redirect or a clicked link, it will first warn about the use of an external browser
+
 ### Bugfixes
 - `<sub>` tags in HTML responses are no longer incorrectly displayed as superscript
 - Variable value resolution no longer fails under some rare conditions